🛠 (c/s): HEAVY changes to admin panel, requiring mini server overhaul

+++ server/models/season_likes.model.sql
@@ -1,7 +1,7 @@
 
 CREATE TABLE IF NOT EXISTS season_likes(
     id SERIAL NOT NULL,
-    season_id INT,
+    season_id REAL,
     user_id INT,
     PRIMARY KEY(id),
     CONSTRAINT fk_season
